% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_config}
\alias{load_config}
\title{Load scilintr configuration from a project root.}
\usage{
load_config(root = ".")
}
\description{
Looks for \code{.scilintr.yml}, \code{analysis_labels.yml}, and
\code{analysis_identifiers.yml} at the project root.
}
\keyword{internal}
