% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_index.R
\name{parse_file_info}
\alias{parse_file_info}
\title{Parse a single file and pull defs, sources, and stage tag.}
\usage{
parse_file_info(path, path_abs)
}
\description{
Parse a single file and pull defs, sources, and stage tag.
}
\keyword{internal}
