% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_index.R
\name{stage_row}
\alias{stage_row}
\title{Stage tag row for the index -- wraps \code{detect_file_stage()}.}
\usage{
stage_row(path, path_abs)
}
\description{
Stage tag row for the index -- wraps \code{detect_file_stage()}.
}
\keyword{internal}
