% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule-R043-unvalidated-config.R
\name{unvalidated_config_linter}
\alias{unvalidated_config_linter}
\title{Flag \code{yaml::read_yaml(...)} / \code{yaml::yaml.load_file(...)} /
\code{jsonlite::fromJSON(...)} results that are later accessed via
\code{$} or \verb{[[ ]]} without an intervening validator call.

[ ]: R:\%20}
\usage{
unvalidated_config_linter()
}
\description{
Track variables whose value comes from one of the loader calls
above. If the variable is then read by \verb{<var>$<key>} or
\verb{<var>[["<key>"]]}, flag the loader call. If the variable is
passed to a \verb{validate_*} or \verb{_schema} function before being
read, no finding fires.
}
\details{
Ported from the Python rule \code{unvalidated-config}.
}
\keyword{internal}
