% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{main}
\alias{main}
\title{Main CLI entry point.}
\usage{
main(args = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{args}{Character vector of command-line arguments.}
}
\description{
Invoked from \code{inst/bin/scilintr} or
\verb{Rscript -e 'scilintr::main()'}.
}
